Finish Flight fallback and completion handling
